fix(stats): 🐛 ensure avgScore serializes as float and validate with Number.isFinite
PostgreSQL ROUND() returns numeric type which pg driver serializes as
string. Add ::float cast to force numeric serialization. On the frontend,
use Number.isFinite(Number(x)) instead of typeof === "number" to handle
both NaN and string-from-pg edge cases.

+++ src/app/api/v1/contests/[assignmentId]/stats/route.ts
@@ -88,7 +88,7 @@ export const GET = createApiHandler({
       )
       SELECT
         COUNT(*)::int AS "submittedCount",
-        COALESCE(ROUND(AVG(ut.total_score), 1), 0) AS "avgScore"
+        COALESCE(ROUND(AVG(ut.total_score), 1), 0)::float AS "avgScore"
       FROM user_totals ut`,
       { assignmentId }
     );
